Final assignment.
I would have liked to have done more, but I'd sort of backed-myself-into-a-corner using the code I already had, and I felt too pushed for time for a rewrite.

Hopefully this will have enough features to pass for something reasonable, though at the very least it should get a pass, we shall see.